requirements.txt
Appium-Python-Client==5.0.0
attrs==25.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
piexif==1.1.3
pillow==10.2.0
pyaes==1.6.1
pyasn1==0.5.1
pyfiglet==1.0.2
PySocks==1.7.1
requests==2.31.0
rsa==4.9
selenium==4.31.0
setuptools==69.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
Telethon==1.34.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.2.1
websocket-client==1.8.0
wheel==0.43.0
wsproto==1.2.0
